requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==4.2.16
django-allauth==0.57.0
django-cors-headers==4.3.1
django-modeltranslation==0.19.15b1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
idna==3.10
mysqlclient==2.2.7
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==79.0.1
sqlparse==0.5.3
urllib3==2.4.0
wheel==0.45.1
